+++ src/js/addon/oneMessage.js
@@ -9,9 +9,6 @@
  */
 (function($) {
     $.fn.bootstrapValidator.addOns.oneMessage = {
-        // Map the field name
-        _map: {},
-
         /**
          * @param {BootstrapValidator} validator The BootstrapValidator instance
          * @param {Object} options The add-on options